requirements.txt
paramiko
psycopg2
neo4j
python-dateutil
numpy
torch
pyqt5
scikit-learn
matplotlib
flask
pandas
plotly
pycallgraph2
psycopg2
sentence-transformers
sqlalchemy
ansi2html
streamlit
psycopg2-binary